requirements.txt
sounddevice
azure-cognitiveservices-speech
streamlit
